// vietsneaker-ui/src/gen/models/AdminOrderSummaryResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AdminOrderSummaryResponse = {
  orderId?: number;
  customerName?: string;
  phone?: string;
  address?: string;
  totalPrice?: number;
  createdAt?: string;
  status?: string;
};